fixed dataset generation implementation bugs;

+++ utils/dirs.py
@@ -17,12 +17,13 @@ def is_empty(dir):
     return len(os.listdir(dir) ) == 0
 
 def clear_dir(dir):
-    for the_file in os.listdir(dir):
-        file_path = os.path.join(dir, the_file)
-        try:
-            if os.path.isfile(file_path):
-                os.unlink(file_path)
-            elif os.path.isdir(file_path):
-                shutil.rmtree(file_path)
-        except Exception as e:
-            print(e)
+    if os.path.isdir(dir):
+        for the_file in os.listdir(dir):
+            file_path = os.path.join(dir, the_file)
+            try:
+                if os.path.isfile(file_path):
+                    os.unlink(file_path)
+                elif os.path.isdir(file_path):
+                    shutil.rmtree(file_path)
+            except Exception as e:
+                print(e)
